requirements.txt
python-telegram-bot
python-dotenv
mega.py